requirements.txt
flask
flask-cors
pinecone-client
openai
httpx
requests
googlemaps
yelpapi
pyowm
beautifulsoup4
requests-html
pydantic
torch
transformers
python-dotenv
firebase-admin
pandas

flask
flask-cors
flask-socketio
python-dotenv
requests
googlemaps
yelpapi
pyowm
beautifulsoup4
requests-html
firebase-admin
pandas
folium

# Optional / used later for scraping, NLP, embeddings
torch
transformers
langchain
pinecone-client
openai
